renamed imports from index to firebase and fixed bootstrap error in Dashboard component
