+++ src/Components/Dashboard.jsx
@@ -1,7 +1,6 @@
 import React from 'react';
 import * as auth from '../Helpers/AuthHelpers';
-import {Link} from 'react-router-dom';
-import firebase from './../Firebase/index'
+import firebase from '../Firebase/firebase'
 import {Navbar, Nav, NavItem} from 'react-bootstrap'
 
 class Dashboard extends React.Component {
@@ -42,28 +41,16 @@ class Dashboard extends React.Component {
                 <Navbar inverse collapseOnSelect>
                     <Navbar.Header>
                         <Navbar.Brand>
-                            <a href="#">Copper Crown</a>
+                            <a href="#">CopperCrown</a>
                         </Navbar.Brand>
                         <Navbar.Toggle />
                     </Navbar.Header>
                     <Navbar.Collapse>
-                        <Nav>
-                            <Navbar.Text pullLeft>
-                                Signed in as: {this.state.name}
-                            </Navbar.Text>
-                        </Nav>
                         <Nav pullRight>
                             <NavItem eventKey={1} onClick={this.logoutUser} href="/login">Logout</NavItem>
                         </Nav>
                     </Navbar.Collapse>
                 </Navbar>
-                <h1>Dashboard Component!</h1>
-                <h2>Welcome, {this.state.name}!</h2>
-                <div>
-                </div>
-                <div>
-                    <Link to="/coppermap">CopperMap</Link>
-                </div>
 
             </div>
         )
